Add auto-generation of a draft prerelease when a `v*` tag is pushed
This change allows our pipeline to auto-run when a new tag is pushed,
ensuring that no manual action is needed other than publishing the
release is needed when we are cutting a release. We currently attach
only the release notes and changelog markdown notes to the release but
there may be more changes that we want later.