Added ability for `repositories` module to point to unreleased versions
By moving the pinned version of a component to be the lowest version we
want to include in our output, this new method can effectively select
all released that have occured after the original configuration
selected. This allows us to coi-opt the same changelog generation to
retrieve all unreleased changes that have a corresponding changelog
entry.