Added ability to output html template-generated files
We now will need to output data to HTML files and because of that we
have to use `html/template` pacakge instead of `text/template` for those
to ensure that we don't directly use escape sequence in the outputs. The
commit for now doesn't change anything from the user perspective but
allows for us to tie this with the rest of the code.