Fixed suite diffing module logic
Previously, old configuration was used as the target for all updates to
the version between the two version but what we really wanted was that
the new config describes which repos we want to use and only update the
AfterVersion field on them based on the older config. This both fixed
the logic and improves the UX since we expect to have the `this` object
be the current/latest config.