Updates parser to organize components by category
In the output of the parser, we want to be able to organize the components
into categories. This will make the output easier to read so that it's not
just a long list of projects, and will hopefully also break up the set of
components into logical subsets of projects.