Added installation instructions to release docs artifacts
This change adds installation instructions to generated `docs-release`
artifacts as they are needed there for our users. This change required a
bit deeper tests changes as we needed to hit more logic branches in the
templates to test properly.

It should be noted that this extra info should ideally be done as a
partial just like it is in the markdown templates but we cannot do that
until https://github.com/cyberark/conjur-oss-helm-chart/issues/50 is
fixed.