Add suite.yml Description to docs release output
We'd like to be able to publish some "What's new" type content with each
suite release, so in this commit I've repurposed the suite Description to
be a Description of what's new in the release. This gets added to the beginning
of the HTML release notes in its own paragraph.