Separate suite iteration logic into own method
In this set of changes we want to add logic to consider the build metadata in
ordering the suite versions. This is inconsistent with the semver standard,
which explicitly excludes build metadata when considering the ordering of
versions; however, the suite is not using strict semver, and the suite versions
do not correspond to semantic versions of the suite, but are tightly linked
to the semantic versioning of the Conjur project itself.

To make it easier to test this logic, the method to determine the suite version
from the suite file semver is broken out into a separate method with its own
unit tests. The "latest release in dir" logic is updated so that it tracks
the highest version's suite iteration, and when the highest version is the same
as the version currently under consideration, the suite iterations are compared
to determine which version is "higher".

This commit also adds a changelog message for this change.