Update strategy for determining tag version
On the v1.11.2+suite.1 release we discovered that the draft-release workflow
was not correctly determining the version from the GITHUB_REFS env var. This
commit updates the logic to determine the version so that this will work
correctly on the next release.

Also removes unneeded version parsing from unit test flow