Fix changelog parsing when repo has main branch
Previously we only checked for if the repo has a release branch, and default to
"master" if not. But we are starting to add projects that use a default "main"
branch, and the logic needs to be updated to accommodate.

This commit adds logic so that if there is no release branch, we check if there
is a "main" branch to use before defaulting to "master". It also adds a test
flow that validates the logic used.