Exclude prereleases from changelog parsing
Under the new release process, repos have a prerelease with version tags that
include a suffix, i.e. v1.2.3-## instead of v1.2.3. This was breaking the OSS
suite release changelog parsing script because it would see the prerelease and
try to match it with a suffix-less tag.

The easiest fix is just to exclude prereleases as we don't need to consider them
when doing changelog parsing for the OSS suite release.